requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
requests>=2.28.0
openpyxl>=3.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Data processing
pyarrow>=8.0.0
fastparquet>=0.8.0
xlrd>=2.0.0
xlsxwriter>=3.0.0

# Database connectors
sqlalchemy-utils>=0.38.0
alembic>=1.8.0

# Logging and monitoring
loguru>=0.6.0
structlog>=22.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
pre-commit>=2.20.0

# Optional: Airflow integration
apache-airflow>=2.5.0

# Optional: Data validation
great-expectations>=0.15.0
pandera>=0.14.0

# Optional: Monitoring
prometheus-client>=0.14.0
